e2e_tests/requirements.txt
pytest==7.4.4
pytest-playwright

ipdb
ipython
